@@ -25,19 +25,14 @@
   },
   "dependencies": {
     "@types/express": ">=4.17.6",
-    "@types/mocha": ">=7.0.2",
-    "@types/node": "^13.13.5",
-    "@types/socket.io": ">=2.1.4",
-    "@types/socket.io-client": ">=1.4.32",
+    "@types/node": "^14.0.1",
+    "@types/socket.io": "^2.1.5",
+    "@types/socket.io-client": "^1.4.33",
     "@types/ssh2": ">=0.5.43",
-    "@typescript-eslint/parser": "^2.30.0",
     "dialog-polyfill": "^0.5.1",
-    "eslint": "5.14.1",
-    "eslint-config-google": "0.12.0",
     "express": "4.16.1",
     "express-winston": "3.1.0",
     "http-auth": "3.2.3",
-    "mocha": "^7.1.2",
     "serve-favicon": "^2.5.0",
     "serve-static": "^1.14.1",
     "socket.io": "2.2.0",
@@ -45,14 +40,12 @@
     "socketio-file-upload": "^0.7.0",
     "ssh2": "0.8.2",
     "ts-node": "^8.9.1",
-    "tslint": "^6.1.2",
-    "typescript": "^3.8.3",
-    "webpack": "^4.43.0",
+    "typescript": "^3.9.2",
     "winston": "^3.2.1"
   },
   "scripts": {
     "tsc": "tsc && tsc -p src/frontend",
-    "build": "npm run lint && npm run tsc && ./node_modules/.bin/webpack",
+    "build": "npm run lint && npm run format && npm run tsc && ./node_modules/.bin/webpack",
     "start": "node dist/index.js",
     "debug-start": "npm run tsc && webpack --config webpack.debug-config.js && npm start",
     "basic": "npm run build && npm start M2 docker",
@@ -65,18 +58,21 @@
     "ssh_singular": "npm run build && npm start singular ssh",
     "forever_singular": "npm run build && forever start -l /home/user/logs/forever.log -o /home/user/logs/out.log -e /home/user/logs/err.log --append dist/index.js singular ssh",
     "stop": "kill `ps ax | grep Server | grep -v monitor | grep node | awk '{print $1}'`",
-    "test": "mocha -r ts-node/register src/tests/**/*.ts --reporter spec --timeout 6000",
-    "lint": "tslint -c tslint.json src/lib/*.ts src/startupConfigs/*.ts --fix && eslint ./public-source/ --fix"
+    "lint": "eslint src/lib/*.ts src/startupConfigs/*.ts src/frontend/*.ts ./public-source/ --fix",
+    "format": "prettier --write src/lib/*.ts src/startupConfigs/*.ts src/frontend/*.ts"
   },
   "repository": {
     "type": "git",
     "url": "git@github.com:pzinn/InteractiveShell.git"
   },
-  "eslintConfig": {
-    "extends": "google",
-    "parser": "@typescript-eslint/parser"
-  },
   "devDependencies": {
-    "webpack-cli": ">=3.2.3"
+    "@typescript-eslint/eslint-plugin": "^2.33.0",
+    "@typescript-eslint/parser": "^2.33.0",
+    "eslint": "^5.16.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.3",
+    "prettier": "^2.0.5",
+    "webpack": "^5.0.0-beta.16",
+    "webpack-cli": "^4.0.0-beta.8"
   }
 }
