@@ -1,8 +1,12 @@
 /* eslint-env browser */
 
-'use strict;';
+"use strict;";
 
-document.addEventListener('DOMContentLoaded', function() {
-  const init = require('../dist/frontend/mathProgram');
-  init();
-}, false);
+document.addEventListener(
+  "DOMContentLoaded",
+  function () {
+    const init = require("../dist/frontend/mathProgram");
+    init();
+  },
+  false
+);
