@@ -1,6 +1,13 @@
 // Emit one of these types via websocket.
-enum SocketEvent {"result", "image", "viewHelp"}
-export {SocketEvent};
+enum SocketEvent {
+  "result",
+  "image",
+  "viewHelp",
+}
+export { SocketEvent };
 
-enum AuthOption {"basic", "none"}
-export {AuthOption};
+enum AuthOption {
+  "basic",
+  "none",
+}
+export { AuthOption };
