@@ -1,8 +1,8 @@
-import {Instance} from "./instance";
+import { Instance } from "./instance";
 
 interface InstanceManager {
-    getNewInstance(next: any);
-    updateLastActiveTime(instance: Instance);
+  getNewInstance(next: any);
+  updateLastActiveTime(instance: Instance);
 }
 
-export {InstanceManager};
+export { InstanceManager };
