@@ -1,15 +1,15 @@
 "use strict;";
 
-import {Client} from "./client";
-import {IClients} from "./client";
+import { Client } from "./client";
+import { IClients } from "./client";
 import clientIdHelper from "./clientId";
 
-import {AuthOption, SocketEvent} from "../lib/enums";
-import {Instance} from "./instance";
-import {InstanceManager} from "./instanceManager";
-import {LocalContainerManager} from "./LocalContainerManager";
-import {SshDockerContainers} from "./sshDockerContainers";
-import {SudoDockerContainers} from "./sudoDockerContainers";
+import { AuthOption, SocketEvent } from "../lib/enums";
+import { Instance } from "./instance";
+import { InstanceManager } from "./instanceManager";
+import { LocalContainerManager } from "./LocalContainerManager";
+import { SshDockerContainers } from "./sshDockerContainers";
+import { SudoDockerContainers } from "./sudoDockerContainers";
 import { AddressInfo } from "net";
 import { emitUrlForUserGeneratedFileToClient } from "./fileDownload";
 
@@ -31,24 +31,24 @@ const webAppTags = require("../frontend/tags");
 import path = require("path");
 let getClientIdFromSocket;
 let serverConfig = {
-  MATH_PROGRAM : undefined,
-  CMD_LOG_FOLDER : undefined,
-  MATH_PROGRAM_COMMAND : undefined,
-  resumeString : undefined,
-  port : undefined,
-  CONTAINERS : undefined,
+  MATH_PROGRAM: undefined,
+  CMD_LOG_FOLDER: undefined,
+  MATH_PROGRAM_COMMAND: undefined,
+  resumeString: undefined,
+  port: undefined,
+  CONTAINERS: undefined,
 };
 let options;
 const staticFolder = path.join(__dirname, "../../public/public");
 let myLogger;
 
-const logExceptOnTest = function(msg: string): void {
+const logExceptOnTest = function (msg: string): void {
   if (process.env.NODE_ENV !== "test") {
     console.log(msg);
   }
 };
 
-const sshCredentials = function(instance: Instance): ssh2.ConnectConfig {
+const sshCredentials = function (instance: Instance): ssh2.ConnectConfig {
   return {
     host: instance.host,
     port: instance.port,
@@ -59,36 +59,36 @@ const sshCredentials = function(instance: Instance): ssh2.ConnectConfig {
 
 const clients: IClients = {};
 
-let totalUsers: number = 0;
+let totalUsers = 0;
 
 let instanceManager: InstanceManager = {
-    getNewInstance(next: any) {
-      //
-    },
-    updateLastActiveTime() {
-      //
-    },
+  getNewInstance(next: any) {
+    //
+  },
+  updateLastActiveTime() {
+    //
+  },
 };
 
-const logClient = function(clientID, str) {
+const logClient = function (clientID, str) {
   if (process.env.NODE_ENV !== "test") {
     logExceptOnTest(clientID + ": " + str);
   }
 };
 
-const userSpecificPath = function(client: Client): string {
+const userSpecificPath = function (client: Client): string {
   return "/" + client.id + "-files/";
 };
 
-const disconnectSocket = function(socket: SocketIO.Socket): void  {
-      try {
-        socket.disconnect();
-      } catch (error) {
-        logExceptOnTest("Failed to disconnect socket: " + error);
-      }
-      };
+const disconnectSocket = function (socket: SocketIO.Socket): void {
+  try {
+    socket.disconnect();
+  } catch (error) {
+    logExceptOnTest("Failed to disconnect socket: " + error);
+  }
+};
 
-const disconnectSockets = function(sockets): void  {
+const disconnectSockets = function (sockets): void {
   for (const socketKey in sockets) {
     if (sockets.hasOwnProperty(socketKey)) {
       const socket: SocketIO.Socket = sockets[socketKey];
@@ -97,16 +97,18 @@ const disconnectSockets = function(sockets): void  {
   }
 };
 
-const deleteClientData = function(client: Client): void {
-  logClient(client.id, "deleting folder " +
-      staticFolder + userSpecificPath(client));
+const deleteClientData = function (client: Client): void {
+  logClient(
+    client.id,
+    "deleting folder " + staticFolder + userSpecificPath(client)
+  );
   try {
     logClient(client.id, "Sending disconnect. ");
     disconnectSockets(clients[client.id].socketArray);
   } catch (error) {
     logClient(client.id, "Socket seems already dead: " + error);
   }
-  fs.rmdir(staticFolder + userSpecificPath(client), function(error) {
+  fs.rmdir(staticFolder + userSpecificPath(client), function (error) {
     if (error) {
       logClient(client.id, "Error deleting user folder: " + error);
     }
@@ -114,7 +116,11 @@ const deleteClientData = function(client: Client): void {
   delete clients[client.id];
 };
 
-const emitDataViaSockets = function(sockets, type: SocketEvent, data: string): void {
+const emitDataViaSockets = function (
+  sockets,
+  type: SocketEvent,
+  data: string
+): void {
   for (const socketKey in sockets) {
     if (sockets.hasOwnProperty(socketKey)) {
       const socket = sockets[socketKey];
@@ -123,104 +129,142 @@ const emitDataViaSockets = function(sockets, type: SocketEvent, data: string): v
   }
 };
 
-const emitDataSafelyViaSocket = function(socket, type: SocketEvent, data: string): void {
+const emitDataSafelyViaSocket = function (
+  socket,
+  type: SocketEvent,
+  data: string
+): void {
   try {
     socket.emit(SocketEvent[type], data);
   } catch (error) {
-    logExceptOnTest("Error while executing socket.emit of type " + SocketEvent[type]);
+    logExceptOnTest(
+      "Error while executing socket.emit of type " + SocketEvent[type]
+    );
   }
 };
 
-const emitDataViaClientSockets = function(client: Client, type: SocketEvent, data) {
+const emitDataViaClientSockets = function (
+  client: Client,
+  type: SocketEvent,
+  data
+) {
   const sockets = client.socketArray;
   emitDataViaSockets(sockets, type, data);
 };
 
-const getInstance = function(client: Client, next) {
+const getInstance = function (client: Client, next) {
   if (client.instance) {
     next(client.instance);
   } else {
     try {
-      instanceManager.getNewInstance(function(err, instance: Instance) {
+      instanceManager.getNewInstance(function (err, instance: Instance) {
         if (err) {
-          emitDataViaClientSockets(client, SocketEvent.result,
+          emitDataViaClientSockets(
+            client,
+            SocketEvent.result,
             "Sorry, there was an error. Please come back later.\n" +
-              err + "\n\n");
+              err +
+              "\n\n"
+          );
           deleteClientData(client);
         } else {
           next(instance);
         }
       });
     } catch (error) {
-      logClient(client.id, "Could not get new instance. Should not drop in here.");
+      logClient(
+        client.id,
+        "Could not get new instance. Should not drop in here."
+      );
     }
   }
 };
 
-export {emitDataViaClientSockets, serverConfig, clients, getInstance, instanceManager, sendDataToClient};
+export {
+  emitDataViaClientSockets,
+  serverConfig,
+  clients,
+  getInstance,
+  instanceManager,
+  sendDataToClient,
+};
 
-const optLogCmdToFile = function(clientId: string, msg: string) {
+const optLogCmdToFile = function (clientId: string, msg: string) {
   if (serverConfig.CMD_LOG_FOLDER) {
-    fs.appendFile(serverConfig.CMD_LOG_FOLDER + "/" + clientId + ".log",
+    fs.appendFile(
+      serverConfig.CMD_LOG_FOLDER + "/" + clientId + ".log",
       msg,
-      function(err) {
+      function (err) {
         if (err) {
           logClient(clientId, "logging msg failed: " + err);
         }
-      });
+      }
+    );
   }
 };
 
-const killNotify = function(client: Client) {
-  return function() {
+const killNotify = function (client: Client) {
+  return function () {
     logClient(client.id, "getting killed.");
     deleteClientData(client);
     optLogCmdToFile(client.id, "Killed.\n");
   };
 };
 
-const spawnMathProgramInSecureContainer = function(client: Client) {
+const spawnMathProgramInSecureContainer = function (client: Client) {
   logClient(client.id, "Spawning new MathProgram process...");
-  getInstance(client, function(instance: Instance) {
+  getInstance(client, function (instance: Instance) {
     instance.killNotify = killNotify(client);
     const connection: ssh2.Client = new ssh2.Client();
-    connection.on("error", function(err) {
-      logClient(client.id, "Error when connecting. " + err +
-        "; Retrying with new instance.");
+    connection.on("error", function (err) {
+      logClient(
+        client.id,
+        "Error when connecting. " + err + "; Retrying with new instance."
+      );
       // Make sure the sanitizer runs.
       try {
         delete client.instance;
         client.saneState = true;
         sanitizeClient(client);
       } catch (instanceDeleteError) {
-        logClient(client.id, "Error when deleting instance: " + instanceDeleteError);
+        logClient(
+          client.id,
+          "Error when deleting instance: " + instanceDeleteError
+        );
         deleteClientData(client);
       }
     });
-    connection.on("ready", function() {
-      client.instance = instance;
-      connection.exec(serverConfig.MATH_PROGRAM_COMMAND,
-        {pty: true},
-        function(err, channel: ssh2.ClientChannel) {
-          if (err) {
-            throw err;
+    connection
+      .on("ready", function () {
+        client.instance = instance;
+        connection.exec(
+          serverConfig.MATH_PROGRAM_COMMAND,
+          { pty: true },
+          function (err, channel: ssh2.ClientChannel) {
+            if (err) {
+              throw err;
+            }
+            optLogCmdToFile(client.id, "Starting.\n");
+            channel.on("close", function () {
+              connection.end();
+            });
+            channel.on("end", function () {
+              channel.close();
+              logClient(
+                client.id,
+                "Channel to Math program ended, closing connection."
+              );
+              connection.end();
+            });
+            attachChannelToClient(client, channel);
           }
-          optLogCmdToFile(client.id, "Starting.\n");
-          channel.on("close", function() {
-            connection.end();
-          });
-          channel.on("end", function() {
-            channel.close();
-            logClient(client.id, "Channel to Math program ended, closing connection.");
-            connection.end();
-          });
-          attachChannelToClient(client, channel);
-        });
-    }).connect(sshCredentials(instance));
+        );
+      })
+      .connect(sshCredentials(instance));
   });
 };
 
-const updateLastActiveTime = function(client: Client) {
+const updateLastActiveTime = function (client: Client) {
   try {
     instanceManager.updateLastActiveTime(client.instance);
   } catch (noInstanceError) {
@@ -229,21 +273,21 @@ const updateLastActiveTime = function(client: Client) {
   }
 };
 
-const addNewSocket = function(client: Client, socket: SocketIO.Socket) {
+const addNewSocket = function (client: Client, socket: SocketIO.Socket) {
   logClient(client.id, "Adding new socket");
   const socketID: string = socket.id;
   client.socketArray[socketID] = socket;
 };
 
-const sendDataToClient = function(client: Client) {
-  return function(dataObject) {
+const sendDataToClient = function (client: Client) {
+  return function (dataObject) {
     const data: string = dataObject.toString();
     if (client.nSockets() === 0) {
       logClient(client.id, "Error, no socket for client.");
       return;
     }
     updateLastActiveTime(client);
-      /*
+    /*
 	const pathPrefix: string = staticFolder + "-" + serverConfig.MATH_PROGRAM;
     const specialUrlEmitter = require("./specialUrlEmitter")(
       pathPrefix,
@@ -263,17 +307,17 @@ const sendDataToClient = function(client: Client) {
       return;
       }
       */
-      myLogger.log({
-	  level: 'info',
-	  message: data,
-	  cat: 'output',
-	  id: client.id
-      });
+    myLogger.log({
+      level: "info",
+      message: data,
+      cat: "output",
+      id: client.id,
+    });
     emitDataViaClientSockets(client, SocketEvent.result, data);
   };
 };
 
-const attachListenersToOutput = function(client: Client) {
+const attachListenersToOutput = function (client: Client) {
   if (client.channel) {
     client.channel
       .removeAllListeners("data")
@@ -281,21 +325,27 @@ const attachListenersToOutput = function(client: Client) {
   }
 };
 
-const attachChannelToClient = function(client: Client, channel: ssh2.ClientChannel){
+const attachChannelToClient = function (
+  client: Client,
+  channel: ssh2.ClientChannel
+) {
   channel.setEncoding("utf8");
   client.channel = channel;
   attachListenersToOutput(client);
-  setTimeout(function() {
+  setTimeout(function () {
     client.saneState = true;
   }, 2000); // Always need a little time before start is done.
 };
 
-const killMathProgram = function(channel: ssh2.ClientChannel, clientID: string) {
+const killMathProgram = function (
+  channel: ssh2.ClientChannel,
+  clientID: string
+) {
   logClient(clientID, "killMathProgramClient.");
   channel.close();
 };
 
-const unhandled = function(request, response) {
+const unhandled = function (request, response) {
   logExceptOnTest("Request for something we don't serve: " + request.url);
   response.writeHead(404, "Request for something we don't serve.");
   response.write("404");
@@ -310,37 +360,42 @@ const getHelp = function(req, res, next) {
     res.redirect(301, 'http://www2.macaulay2.com/Macaulay2/doc/Macaulay2/share/doc/Macaulay2'+req.path);
     }
 */
-const getHelp = function(req, res) {
-    res.sendFile(staticFolder + "-" + serverConfig.MATH_PROGRAM + "/help.html");
-}
+const getHelp = function (req, res) {
+  res.sendFile(staticFolder + "-" + serverConfig.MATH_PROGRAM + "/help.html");
+};
 
-const initializeServer = function() {
+const initializeServer = function () {
   const favicon = require("serve-favicon");
   const serveStatic = require("serve-static");
   const winston = require("winston");
   const expressWinston = require("express-winston");
 
-    const webAppTagsRegExp = new RegExp("(" + Object.values(webAppTags).join("|") + ")","g");
-    const prettyTags = {};
-    const temp = Object.entries(webAppTags);
-    //    for (const a of temp) prettyTags[a[1] as any]="\u001b[32m<"+a[0]+">\u001b[39m";
-        for (const a of temp) prettyTags[a[1] as any]="<"+a[0]+">";
+  const webAppTagsRegExp = new RegExp(
+    "(" + Object.values(webAppTags).join("|") + ")",
+    "g"
+  );
+  const prettyTags = {};
+  const temp = Object.entries(webAppTags);
+  //    for (const a of temp) prettyTags[a[1] as any]="\u001b[32m<"+a[0]+">\u001b[39m";
+  for (const a of temp) prettyTags[a[1] as any] = "<" + a[0] + ">";
 
-    const myFormat = winston.format.printf((info)=>{
-	const msg = info.message.replace(webAppTagsRegExp, function(match,token) { return prettyTags[token]; });
-	// 	return `\u001b[34m${info.cat} to ${info.id}\u001b[39m\n${msg}`;
-	return `${info.cat} to ${info.id}\n${msg}`;
+  const myFormat = winston.format.printf((info) => {
+    const msg = info.message.replace(webAppTagsRegExp, function (match, token) {
+      return prettyTags[token];
     });
+    // 	return `\u001b[34m${info.cat} to ${info.id}\u001b[39m\n${msg}`;
+    return `${info.cat} to ${info.id}\n${msg}`;
+  });
 
-
-  myLogger=winston.createLogger({ // custom logger (as opposed to express winston)
-      level: 'info',
-      format: myFormat,
-	transports: [
-	    new winston.transports.File({
-		filename: "winston.log"
-	    })
-	],
+  myLogger = winston.createLogger({
+    // custom logger (as opposed to express winston)
+    level: "info",
+    format: myFormat,
+    transports: [
+      new winston.transports.File({
+        filename: "winston.log",
+      }),
+    ],
   });
 
   const loggerSettings = {
@@ -349,22 +404,23 @@ const initializeServer = function() {
         level: "error",
         json: true,
         colorize: true,
-      })
+      }),
     ],
   };
 
   const prefix: string = staticFolder + "-" + serverConfig.MATH_PROGRAM + "/";
   const getList: reader.GetListFunction = reader.tutorialReader(prefix, fs);
   const admin = require("./admin")(clients, -1, serverConfig.MATH_PROGRAM);
-  app.use(favicon(staticFolder + "-" +
-      serverConfig.MATH_PROGRAM + "/favicon.ico"));
-  app.use(SocketIOFileUpload.router);  
+  app.use(
+    favicon(staticFolder + "-" + serverConfig.MATH_PROGRAM + "/favicon.ico")
+  );
+  app.use(SocketIOFileUpload.router);
   // to obtain the raw files
-  app.use("/force/usr/share/",serveStatic(staticFolder + "-share"));
+  app.use("/force/usr/share/", serveStatic(staticFolder + "-share"));
   // otherwise html's get processed
-  app.get(/\/usr\/share\/.+\.html/,getHelp);
+  app.get(/\/usr\/share\/.+\.html/, getHelp);
   // rest is fine
-  app.use("/usr/share/",serveStatic(staticFolder + "-share"));
+  app.use("/usr/share/", serveStatic(staticFolder + "-share"));
   app.use(serveStatic(staticFolder + "-" + serverConfig.MATH_PROGRAM));
   app.use(serveStatic(staticFolder + "-common"));
   app.use(expressWinston.logger(loggerSettings));
@@ -373,41 +429,41 @@ const initializeServer = function() {
   app.use(unhandled);
 };
 
-const clientExistenceCheck = function(clientId: string, socket): Client {
+const clientExistenceCheck = function (clientId: string, socket): Client {
   logExceptOnTest("Checking existence of client with id " + clientId);
   if (!clients[clientId]) {
     clients[clientId] = new Client(clientId);
     totalUsers += 1;
   } else {
-    emitDataSafelyViaSocket(socket,
-       SocketEvent.result,
-      "Session resumed.\n" + serverConfig.resumeString);
+    emitDataSafelyViaSocket(
+      socket,
+      SocketEvent.result,
+      "Session resumed.\n" + serverConfig.resumeString
+    );
   }
   return clients[clientId];
 };
 
-const sanitizeClient = function(client: Client) {
+const sanitizeClient = function (client: Client) {
   if (!client.saneState) {
     logClient(client.id, "Is already being sanitized.");
   }
   client.saneState = false;
 
-  if (!client.channel ||
-      !client.channel.writable ||
-      !client.instance) {
+  if (!client.channel || !client.channel.writable || !client.instance) {
     spawnMathProgramInSecureContainer(client);
   } else {
     logClient(client.id, "Has mathProgram instance.");
     client.saneState = true;
   }
   // Avoid stuck sanitizer.
-  setTimeout(function() {
+  setTimeout(function () {
     client.saneState = true;
   }, 2000);
 };
 
-const writeMsgOnStream = function(client: Client, msg: string) {
-  client.channel.stdin.write(msg, function(err) {
+const writeMsgOnStream = function (client: Client, msg: string) {
+  client.channel.stdin.write(msg, function (err) {
     if (err) {
       logClient(client.id, "write failed: " + err);
       sanitizeClient(client);
@@ -416,18 +472,17 @@ const writeMsgOnStream = function(client: Client, msg: string) {
   });
 };
 
-const checkAndWrite = function(client: Client, msg: string) {
-  if (!client.channel ||
-      !client.channel.writable) {
+const checkAndWrite = function (client: Client, msg: string) {
+  if (!client.channel || !client.channel.writable) {
     sanitizeClient(client);
   } else {
     writeMsgOnStream(client, msg);
   }
 };
 
-const checkClientSanity = function(client: Client) {
+const checkClientSanity = function (client: Client) {
   logClient(client.id, "Checking sanity");
-  return new Promise(function(resolve, reject) {
+  return new Promise(function (resolve, reject) {
     if (client.saneState) {
       logClient(client.id, "Is sane");
       resolve();
@@ -435,15 +490,15 @@ const checkClientSanity = function(client: Client) {
       logClient(client.id, "Not accepting events.");
       reject();
     }
-  }).catch(()=>{
-      // empty
+  }).catch(() => {
+    // empty
   });
 };
 
-const socketInputAction = function(socket, client: Client) {
-  return function(msg: string) {
+const socketInputAction = function (socket, client: Client) {
+  return function (msg: string) {
     logClient(client.id, "Receiving input");
-    checkClientSanity(client).then(function() {
+    checkClientSanity(client).then(function () {
       setCookieOnSocket(socket, client.id);
       updateLastActiveTime(client);
       checkAndWrite(client, msg);
@@ -451,50 +506,53 @@ const socketInputAction = function(socket, client: Client) {
   };
 };
 
-const socketResetAction = function(client: Client) {
-  return function() {
+const socketResetAction = function (client: Client) {
+  return function () {
     optLogCmdToFile(client.id, "Resetting.\n");
     logClient(client.id, "Received reset.");
-    checkClientSanity(client).then(function() {
+    checkClientSanity(client).then(function () {
       if (client.channel) {
         killMathProgram(client.channel, client.id);
-	spawnMathProgramInSecureContainer(client);
+        spawnMathProgramInSecureContainer(client);
       }
       sanitizeClient(client);
     });
   };
 };
 
-const socketDownloadAction = function(socket, client: Client) {
-    const pathPrefix: string = staticFolder + "-" + serverConfig.MATH_PROGRAM;
-    return function(msg: string) {
-      emitUrlForUserGeneratedFileToClient(
-        client,
-        msg,
-        pathPrefix,
-          userSpecificPath(client),
-        sshCredentials,
-        logExceptOnTest,
-        emitDataViaSockets);
-    };
+const socketDownloadAction = function (socket, client: Client) {
+  const pathPrefix: string = staticFolder + "-" + serverConfig.MATH_PROGRAM;
+  return function (msg: string) {
+    emitUrlForUserGeneratedFileToClient(
+      client,
+      msg,
+      pathPrefix,
+      userSpecificPath(client),
+      sshCredentials,
+      logExceptOnTest,
+      emitDataViaSockets
+    );
+  };
 };
 
 const sevenDays = 7 * 86409000;
 
-const initializeClientId = function(socket): string{
+const initializeClientId = function (socket): string {
   const clientID = clientIdHelper(clients, logExceptOnTest).getNewId();
   setCookieOnSocket(socket, clientID);
   return clientID;
 };
 
-const setCookieOnSocket = function(socket, clientID: string): void{
+const setCookieOnSocket = function (socket, clientID: string): void {
   const expDate = new Date(new Date().getTime() + sevenDays);
-  const sessionCookie = Cookie.serialize(options.cookieName, clientID, {expires: expDate});
+  const sessionCookie = Cookie.serialize(options.cookieName, clientID, {
+    expires: expDate,
+  });
   socket.emit("cookie", sessionCookie);
 };
 
-const listen = function() {
-  io.on("connection", function(socket: SocketIO.Socket) {
+const listen = function () {
+  io.on("connection", function (socket: SocketIO.Socket) {
     logExceptOnTest("Incoming new connection!");
     let clientId: string = getClientIdFromSocket(socket);
     if (typeof clientId === "undefined") {
@@ -509,21 +567,21 @@ const listen = function() {
     const client = clientExistenceCheck(clientId, socket);
     sanitizeClient(client);
     addNewSocket(client, socket);
-    const fileUpload = require("./fileUpload")(
-      logExceptOnTest,
-      sshCredentials);
+    const fileUpload = require("./fileUpload")(logExceptOnTest, sshCredentials);
     fileUpload.attachUploadListenerToSocket(client, socket);
     socket.on("input", socketInputAction(socket, client));
     socket.on("reset", socketResetAction(client));
-      socket.on("download", socketDownloadAction(socket,client));
+    socket.on("download", socketDownloadAction(socket, client));
   });
 
   const listener = http.listen(serverConfig.port);
-    logExceptOnTest("Server running on " + (listener.address() as AddressInfo).port);
+  logExceptOnTest(
+    "Server running on " + (listener.address() as AddressInfo).port
+  );
   return listener;
 };
 
-const authorizeIfNecessary = function(authOption: AuthOption) {
+const authorizeIfNecessary = function (authOption: AuthOption) {
   if (authOption === AuthOption.basic) {
     const auth = require("http-auth");
     const basic = auth.basic({
@@ -531,7 +589,7 @@ const authorizeIfNecessary = function(authOption: AuthOption) {
       file: path.join(__dirname, "/../../../public/users.htpasswd"),
     });
     app.use(auth.connect(basic));
-    return function(socket: SocketIO.Socket) {
+    return function (socket: SocketIO.Socket) {
       try {
         return socket.request.headers.authorization.substring(6);
       } catch (error) {
@@ -539,7 +597,7 @@ const authorizeIfNecessary = function(authOption: AuthOption) {
       }
     };
   }
-  return function(socket: SocketIO.Socket) {
+  return function (socket: SocketIO.Socket) {
     const rawCookies = socket.request.headers.cookie;
     if (typeof rawCookies === "undefined") {
       // Sometimes there are no cookies
@@ -551,7 +609,7 @@ const authorizeIfNecessary = function(authOption: AuthOption) {
   };
 };
 
-const MathServer = function(o) {
+const MathServer = function (o) {
   options = o;
   serverConfig = options.serverConfig;
 
@@ -564,7 +622,11 @@ const MathServer = function(o) {
   const resources = options.perContainerResources;
   const guestInstance = options.startInstance;
   const hostConfig = options.hostConfig;
-  instanceManager = serverConfig.CONTAINERS(resources, hostConfig, guestInstance);
+  instanceManager = serverConfig.CONTAINERS(
+    resources,
+    hostConfig,
+    guestInstance
+  );
 
   initializeServer();
 
