@@ -1,8 +1,9 @@
 module.exports = {
-  entry: "./public-source/index.js",
-  output: {
-    filename: "public/public-common/index.js"
-  },
-  mode: 'development',
-  devtool: "inline-module-source-map"
+    entry: "./public-source/index.js",
+    output: {
+	path: __dirname, // otherwise is put in "dist/"
+	filename: "public/public-common/index.js"
+    },
+    mode: 'development',
+    devtool: "inline-source-map"
 };
